kuring-80 [수정] 도메인 객체를 data:domain 모듈로 옮김
