+++ data/domain/src/main/java/com/ku_stacks/ku_ring/domain/Push.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.data.model
+package com.ku_stacks.ku_ring.domain
 
 data class Push(
     val articleId: String,
